[Feat] Divide ViewModel
